update the icon image update and delete

+++ app/Services/AdminAuthCMS/ServiceService.php
@@ -176,18 +176,33 @@ class ServiceService
                 'title' => $request->image_title ?? $service->image->title,
             ]);
         }
-        if (Storage::disk('public')->exists($service->icon_path)) {
-            Storage::disk('public')->delete($service->icon_path);
-        }
-        $icon_path = $this->uploadImage($request, 'services/icon', 'icon');
+        $iconPath = $service->icon_path;
 
+        if ($request->hasFile('icon')) {
+
+            if ($service->icon_path && Storage::disk('public')->exists($service->icon_path)) {
+                Storage::disk('public')->delete($service->icon_path);
+            }
+
+            $uploadIcon = $this->uploadImage($request, 'services/icon', 'icon');
+
+            if (! $uploadIcon['success']) {
+                return [
+                    'data' => null,
+                    'message' => $uploadIcon['message'],
+                    'code' => 400,
+                ];
+            }
+
+            $iconPath = $uploadIcon['data'];
+        }
         // Update the service's main data
         $service->update([
             'title' => $request->title ?? $service->title,
             'description' => $request->description ?? $service->description,
             'content' => $request->content ?? $service->content,
             'featured' => $request->featured ?? 0,
-            'icon_path' => $icon_path['data'],
+            'icon_path' => $iconPath,
         ]);
 
         // Return the updated service with its image data
@@ -238,8 +253,10 @@ class ServiceService
                 'code' => 404,
             ];
         }
-        if (Storage::disk('public')->exists($service->icon_path)) {
-            Storage::disk('public')->delete($service->icon_path);
+        if ($service->icon_path) {
+            if (Storage::disk('public')->exists($service->icon_path)) {
+                Storage::disk('public')->delete($service->icon_path);
+            }
         }
         $service->delete();
